# docs/swagger.yaml
basePath: /api
definitions:
...
  models.BatchSaveRequest:
    properties:
      deletes:
        items:
          type: integer
        type: array
      updates:
        items:
          $ref: '#/definitions/models.RouteSegment'
        type: array
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.Location:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      identifier:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      maximumAircraftPerHour:
        type: integer
      routeSegmentId:
        type: integer
      sortOrder:
        type: integer
    type: object
...
  models.RouteSegment:
    properties:
      color:
        type: string
      enabled:
        type: boolean
      eventId:
        type: integer
      facilities:
        type: string
      id:
        type: integer
      identifier:
        type: string
      locations:
        items:
          $ref: '#/definitions/models.Location'
        type: array
      maximumAircraftPerHour:
        type: integer
      providedFacilityProgression:
        items:
          $ref: '#/definitions/models.Sector'
        type: array
      routeRevision:
        type: integer
      routeSegmentGroup:
        type: string
      routeString:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.RouteSegmentTag'
        type: array
    type: object
  models.RouteSegmentTag:
    properties:
      id:
        type: integer
      routeSegmentId:
        type: integer
      tag:
        type: string
    type: object
  models.Sector:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      identifier:
        type: string
      maximumAircraftPerHour:
        type: integer
      sectorBoundaries:
        items:
          $ref: '#/definitions/models.SectorBoundary'
        type: array
    type: object
  models.SectorBoundary:
    properties:
      coordinates:
        items:
          $ref: '#/definitions/models.SectorBoundaryCoordinate'
        type: array
      id:
        type: integer
      maxLatitude:
        type: number
      maxLongitude:
        type: number
      minLatitude:
        type: number
      minLongitude:
        type: number
      sectorId:
        type: integer
    type: object
  models.SectorBoundaryCoordinate:
    properties:
      id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      sectorBoundaryId:
        type: integer
      sortOrder:
        type: integer
    type: object
...
  models.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Central data API for CTP (Cross the Pond) event management. All /api/*
    endpoints require an X-API-Key header.
  title: CTP API
  version: 1.0.0
paths:
...
  /events/{eventId}/route-segments:
    get:
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RouteSegment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List route segments for an event
      tags:
      - route-segments
  /events/{eventId}/sectors:
    get:
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Sector'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List sectors for an event
      tags:
      - sectors
    post:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: integer
      - description: Sector
        in: body
        name: sector
        required: true
        schema:
          $ref: '#/definitions/models.Sector'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Sector'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create sector
      tags:
      - sectors
...
  /route-segments:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RouteSegment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all route segments
      tags:
      - route-segments
    post:
      consumes:
      - application/json
      parameters:
      - description: Route segment
        in: body
        name: segment
        required: true
        schema:
          $ref: '#/definitions/models.RouteSegment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.RouteSegment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create route segment
      tags:
      - route-segments
  /route-segments/{id}:
    delete:
      parameters:
      - description: Route segment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete route segment
      tags:
      - route-segments
    put:
      consumes:
      - application/json
      parameters:
      - description: Route segment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Route segment fields to update
        in: body
        name: segment
        required: true
        schema:
          $ref: '#/definitions/models.RouteSegment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RouteSegment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update route segment
      tags:
      - route-segments
  /route-segments/save:
    post:
      consumes:
      - application/json
      parameters:
      - description: 'Batch payload: updates (array of RouteSegment) and deletes (array
          of IDs)'
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.BatchSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'success: true'
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Batch create/update/delete route segments in a single transaction
      tags:
      - route-segments
  /sectors/{id}:
    delete:
      parameters:
      - description: Sector ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete sector
      tags:
      - sectors
    put:
      consumes:
      - application/json
      parameters:
      - description: Sector ID
        in: path
        name: id
        required: true
        type: integer
      - description: Sector fields to update
        in: body
        name: sector
        required: true
        schema:
          $ref: '#/definitions/models.Sector'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sector'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update sector
      tags:
      - sectors
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"